foodlist state fixed
